Use appropriate method naming of private methods
Also, move the use; methods in tests to use_ok.

The tests are really really far from passing at this point.  Tons
of tests and libs aren't there, but that'll be fixed eventually.

One problem that I caught was that when I changed a private method
in the Thingiverse class, it broke a ton of other classes.  This
is because they are all referencing the private method.  I'm going
to have to fix this, probably via dependency injection, in a future
pull request.