Thingiverse conversion to Thingiverse::Object
I've totally hacked the thingiverse injection into the List classes
at the moment as someone else is working on the List base class.

This allows for thingiverse to be passed into the constructor for
them and then passed onto the children.

Unfortunately, this commit is far larger than I'd like, but I
didn't really find good stoppage points for commits before now.